+++ components/ledger/LedgerDialog.tsx
@@ -26,12 +26,16 @@ import DepositSuccess from './proposal/deposit/Success';
 import VoteConfirm from './proposal/vote/Confirm';
 import VoteSuccess from './proposal/vote/Success';
 import Vote from './proposal/vote/Vote';
+import ActivateValidatorGroupVote from './validatorGroup/activate/Activate';
+import ConfirmActivateValidatorGroupVote from './validatorGroup/activate/Confirm';
+import SuccessActivateValidatorGroupVote from './validatorGroup/activate/Success';
 import ConfirmRevokeValidatorGroup from './validatorGroup/revoke/Confirm';
 import RevokeValidatorGroup from './validatorGroup/revoke/Revoke';
 import SuccessRevokeValidatorGroup from './validatorGroup/revoke/Success';
 import ConfirmVoteValidatorGroup from './validatorGroup/vote/Confirm';
 import SuccessVoteValidatorGroup from './validatorGroup/vote/Success';
 import VoteValidatorGroup from './validatorGroup/vote/Vote';
+
 const useStyles = makeStyles((theme: Theme) =>
     createStyles({
         root: {
@@ -44,11 +48,21 @@ const useStyles = makeStyles((theme: Theme) =>
             [theme.breakpoints.down('sm')]: {
                 width: '7.5rem'
             },
-            width: '9.5rem',
+            width: '20rem',
             padding: '0.5rem',
             textTransform: 'none',
             border: 'solid thin',
             margin: '0.3rem 1rem 0.2rem 0'
+        },
+
+        activateButtonLabel: {
+            justifyContent: 'center',
+            padding: '0.5rem',
+            textTransform: 'none',
+            width: '7.5rem',
+            color: 'rgba(255, 255, 255, 1)',
+            backgroundColor: 'rgba(8, 178, 112, 0.95)',
+            '&:hover, &.Mui-focusVisible': { backgroundColor: 'rgba(8, 178, 112, 0.70)' }
         },
 
         gridContainer: {
@@ -70,7 +84,9 @@ const useStyles = makeStyles((theme: Theme) =>
         errorMessage: {
             color: 'red',
             textAlign: 'center',
-            paddingBottom: '1rem'
+            paddingBottom: '2rem'
+            // wordBreak: 'break-all',
+            // lineBreak: 'anywhere'
         },
 
         circularProgress: {
@@ -110,7 +126,8 @@ const useStyles = makeStyles((theme: Theme) =>
             display: 'block',
             textAlign: 'center',
             paddingTop: '0.5rem',
-            paddingBottom: '0.7rem'
+            paddingBottom: '0.7rem',
+            paddingLeft: '2.5rem'
         },
 
         voteNoButton: {
@@ -209,6 +226,8 @@ type LedgerDialogProps = {
     validatorGroup?: string;
 };
 
+type CheckIfAccountProps = { address: string };
+
 const LedgerDialog = ({
     buttonLabel,
     action,
@@ -228,6 +247,7 @@ const LedgerDialog = ({
     const [ledgerErrorMessage, setLedgerErrorMessage] = React.useState('');
     const [ledgerLoading, setLedgerLoading] = React.useState(false);
     const [dialogError, setDialogError] = useGlobalState('dialogError');
+    const [showCircuralProgress, setShowCircuralProgress] = React.useState(false);
     // const [dialogErrorMessage, setDialogErrorMessage] = useGlobalState('dialogErrorMessage');
     const [tabNumber, setTabNumber] = React.useState(0);
     const [showControlButtons, setShowControlButtons] = React.useState(false);
@@ -238,6 +258,7 @@ const LedgerDialog = ({
         proposalDescription || ''
     );
     const [vote, setVote] = React.useState('');
+    const [hash, setHash] = React.useState('');
 
     const address = currentUser;
 
@@ -267,7 +288,7 @@ const LedgerDialog = ({
                     case 1:
                         return <LockGoldConfirm amount={amount} />;
                     case 2:
-                        return <LockGoldSuccess />;
+                        return <LockGoldSuccess txHash={hash} />;
                     default:
                         return null;
                 }
@@ -291,7 +312,7 @@ const LedgerDialog = ({
                     case 1:
                         return <UnlockGoldConfirm amount={amount} />;
                     case 2:
-                        return <UnlockGoldSuccess />;
+                        return <UnlockGoldSuccess txHash={hash} />;
                     default:
                         return null;
                 }
@@ -310,7 +331,7 @@ const LedgerDialog = ({
                             />
                         );
                     case 2:
-                        return <DepositSuccess />;
+                        return <DepositSuccess txHash={hash} />;
                     default:
                         return null;
                 }
@@ -336,7 +357,7 @@ const LedgerDialog = ({
                             />
                         );
                     case 2:
-                        return <VoteSuccess />;
+                        return <VoteSuccess txHash={hash} />;
                     default:
                         return null;
                 }
@@ -366,7 +387,7 @@ const LedgerDialog = ({
                             />
                         );
                     case 2:
-                        return <SuccessVoteValidatorGroup />;
+                        return <SuccessVoteValidatorGroup txHash={hash} />;
                     default:
                         return null;
                 }
@@ -396,20 +417,73 @@ const LedgerDialog = ({
                             />
                         );
                     case 2:
-                        return <SuccessRevokeValidatorGroup />;
+                        return <SuccessRevokeValidatorGroup txHash={hash} />;
+                    default:
+                        return null;
+                }
+            case 'ValidatorGroupActivateVotes':
+                switch (tabNum) {
+                    case 0:
+                        return (
+                            <ActivateValidatorGroupVote
+                                isLoading={isLoading}
+                                validatorGroup={validatorGroup}
+                            />
+                        );
+                    case 1:
+                        return (
+                            <ConfirmActivateValidatorGroupVote validatorGroup={validatorGroup} />
+                        );
+                    case 2:
+                        return <SuccessActivateValidatorGroupVote txHash={hash} />;
                     default:
                         return null;
                 }
         }
     };
 
+    const checkIfAccount = async ({ address }: CheckIfAccountProps) => {
+        const accountAddress = { address: address };
+        try {
+            const isAccount = await Ledger.isAccount(accountAddress);
+            if (isAccount === true) {
+                return true;
+            } else {
+                setLedgerLoading(true);
+                setLedgerErrorMessage(
+                    `Can't find an account with address ${address} on the chain. Please accept the connection and sign the transaction on your Ledger deivce to create an account. `
+                );
+                setIsLoading(true);
+                const createAccount = await Ledger.createAccount(accountAddress);
+                if (createAccount === true) {
+                    return true;
+                } else {
+                    return false;
+                }
+            }
+        } catch (e) {
+            setLedgerError(true);
+            setLedgerLoading(true);
+            setLedgerErrorMessage(Ledger.checkLedgerErrors(e.message));
+        }
+    };
+
     const handleLock = async () => {
         try {
-            setLedgerLoading(true);
-            const from = currentUser;
-            const lockObject = { amount, from };
-            await Ledger.lockCelo(lockObject);
-            setLedgerLoading(false);
+            const address = { address: currentUser };
+            //check if account exist
+            if (checkIfAccount(address)) {
+                setShowCircuralProgress(true);
+                const from = currentUser;
+                const lockObject = { amount, from };
+                const lockCelo = await Ledger.lockCelo(lockObject);
+                setShowCircuralProgress(false);
+                if (lockCelo && lockCelo.status === true) {
+                    setTabNumber(2);
+                    setLedgerLoading(false);
+                    setHash(lockCelo.blockHash);
+                }
+            }
         } catch (e) {
             setLedgerError(true);
             setLedgerLoading(true);
@@ -419,11 +493,20 @@ const LedgerDialog = ({
 
     const handleUnlock = async () => {
         try {
-            setLedgerLoading(true);
-            const from = currentUser;
-            const unlockObject = { amount, from };
-            await Ledger.unlockCelo(unlockObject);
-            setLedgerLoading(false);
+            // setLedgerLoading(true);
+            const address = { address: currentUser };
+            if (checkIfAccount(address)) {
+                setShowCircuralProgress(true);
+                const from = currentUser;
+                const unlockObject = { amount, from };
+                const unlockCelo = await Ledger.unlockCelo(unlockObject);
+                setShowCircuralProgress(false);
+                if (unlockCelo && unlockCelo.status === true) {
+                    setTabNumber(2);
+                    setLedgerLoading(false);
+                    setHash(unlockCelo.blockHash);
+                }
+            }
         } catch (e) {
             setLedgerError(true);
             setLedgerLoading(true);
@@ -447,13 +530,16 @@ const LedgerDialog = ({
 
     const handleProposalVote = async () => {
         try {
-            setLedgerLoading(true);
-            const from = currentUser;
-            const vote = '';
-            const proposalNumber = getProposalNumber;
-            const voteObject = { proposalNumber, from, vote };
-            await Ledger.voteProposal(voteObject);
-            setLedgerLoading(false);
+            const address = { address: currentUser };
+            if (checkIfAccount(address)) {
+                setLedgerLoading(true);
+                const from = currentUser;
+                const vote = '';
+                const proposalNumber = getProposalNumber;
+                const voteObject = { proposalNumber, from, vote };
+                await Ledger.voteProposal(voteObject);
+                setLedgerLoading(false);
+            }
         } catch (e) {
             setLedgerError(true);
             setLedgerLoading(true);
@@ -463,12 +549,21 @@ const LedgerDialog = ({
 
     const handleValidatorGroupVote = async () => {
         try {
-            setLedgerLoading(true);
-            const from = currentUser;
-            const group = validatorGroup ? validatorGroup : '';
-            const unlockObject = { amount, from, group };
-            await Ledger.voteValidatorGroup(unlockObject);
-            setLedgerLoading(false);
+            const address = { address: currentUser };
+            if (checkIfAccount(address)) {
+                setShowCircuralProgress(true);
+                const from = currentUser;
+                const group = validatorGroup ? validatorGroup : '';
+                const groupVote = { amount, from, group };
+                const voteValidatorGroup = await Ledger.voteValidatorGroup(groupVote);
+
+                setShowCircuralProgress(false);
+                if (voteValidatorGroup && voteValidatorGroup.status === true) {
+                    setTabNumber(2);
+                    setLedgerLoading(false);
+                    setHash(voteValidatorGroup.blockHash);
+                }
+            }
         } catch (e) {
             setLedgerError(true);
             setLedgerLoading(true);
@@ -478,13 +573,48 @@ const LedgerDialog = ({
 
     const handleRevokeValidatorGroupVote = async () => {
         try {
+            const address = { address: currentUser };
+            if (checkIfAccount(address)) {
+                setShowCircuralProgress(true);
+                const account = currentUser;
+                const group = validatorGroup ? validatorGroup : '';
+                const revokeObject = { amount, account, group };
+                const revokeValidatorGroupVote = await Ledger.revokeValidatorGroupVote(
+                    revokeObject
+                );
+                setShowCircuralProgress(false);
+                if (revokeValidatorGroupVote && revokeValidatorGroupVote.status === true) {
+                    setTabNumber(2);
+                    setLedgerLoading(false);
+                    setHash(revokeValidatorGroupVote.blockHash);
+                }
+            }
+        } catch (e) {
+            setLedgerError(true);
             setLedgerLoading(true);
-            //still need to obtain the address of the validator/validator group
-            const account = '';
-            const group = validatorGroup ? validatorGroup : '';
-            const unlockObject = { amount, account, group };
-            await Ledger.revokeValidatorGroupVote(unlockObject);
-            setLedgerLoading(false);
+            setLedgerErrorMessage(Ledger.checkLedgerErrors(e.message));
+        }
+    };
+
+    const handleValidatorGroupActivateVotes = async () => {
+        try {
+            const address = { address: currentUser };
+            if (checkIfAccount(address)) {
+                setShowCircuralProgress(true);
+                const activateVotesObject = {
+                    address: currentUser,
+                    validatorGroupAddress: validatorGroup
+                };
+                const activateValidatorGroupVote = await Ledger.activateVaidatorGroupVotes(
+                    activateVotesObject
+                );
+                setShowCircuralProgress(false);
+                if (activateValidatorGroupVote && activateValidatorGroupVote.status === true) {
+                    setTabNumber(2);
+                    setLedgerLoading(false);
+                    setHash(activateValidatorGroupVote.blockHash);
+                }
+            }
         } catch (e) {
             setLedgerError(true);
             setLedgerLoading(true);
@@ -527,17 +657,24 @@ const LedgerDialog = ({
                         return null as any;
                 }
                 break;
-            case 'Validator Group Vote':
+            case 'ValidatorGroupVote':
                 switch (tabNumber) {
                     case 0:
                         return handleValidatorGroupVote();
                     default:
                         return null as any;
                 }
-            case 'Validator Group Revoke':
+            case 'ValidatorGroupRevoke':
                 switch (tabNumber) {
                     case 0:
                         return handleRevokeValidatorGroupVote();
+                    default:
+                        return null as any;
+                }
+            case 'ValidatorGroupActivateVotes':
+                switch (tabNumber) {
+                    case 0:
+                        return handleValidatorGroupActivateVotes();
                     default:
                         return null as any;
                 }
@@ -596,7 +733,7 @@ const LedgerDialog = ({
                 }
                 try {
                     const ver = await Ledger.getCeloAppVersion();
-                    setDialogError(true);
+                    // setDialogError(true);
                 } catch (e) {
                     setLedgerError(true);
                     setLedgerErrorMessage(Ledger.checkLedgerErrors(e.message));
@@ -641,9 +778,17 @@ const LedgerDialog = ({
                             variant="outlined"
                             color="secondary"
                             onClick={handleClick}
-                            style={{ width: action === 'ValidatorGroupVote' ? '21rem' : '7.5rem' }}
-                            className={classes.buttonLabel}>
-                            <Typography variant="body1">{buttonLabel}</Typography>
+                            className={
+                                action === 'ValidatorGroupActivateVotes'
+                                    ? classes.activateButtonLabel
+                                    : classes.buttonLabel
+                            }>
+                            <Typography
+                                variant={
+                                    action === 'ValidatorGroupActivateVotes' ? 'caption' : 'body1'
+                                }>
+                                {buttonLabel}
+                            </Typography>
                         </Button>
                     </div>
                 </Grid>
@@ -657,7 +802,7 @@ const LedgerDialog = ({
                 maxWidth="sm">
                 <DialogTitle id="ledger-title" className={classes.dialogTitle}>
                     <Grid container className={classes.item}>
-                        {tabNumber > 0 ? (
+                        {tabNumber === 1 ? (
                             <Grid item xs={1}>
                                 <IconButton
                                     aria-label="Return"
@@ -688,7 +833,7 @@ const LedgerDialog = ({
                 </DialogTitle>
                 <div className={classes.ledgerDialogPopup}>
                     {dialogTab(tabNumber)}
-                    {ledgerLoading ? (
+                    {ledgerLoading || showCircuralProgress ? (
                         <Grid item xs={12} className={classes.circularProgress}>
                             <CircularProgress color="secondary" />
                         </Grid>