+++ components/ledger/validatorGroup/activate/Confirm.tsx
@@ -1,7 +1,6 @@
 import DialogContent from '@material-ui/core/DialogContent';
 import DialogContentText from '@material-ui/core/DialogContentText';
 import Divider from '@material-ui/core/Divider';
-//import Link from "../../Ledger.tsx";
 import Grid from '@material-ui/core/Grid';
 import { makeStyles } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
@@ -13,8 +12,7 @@ const useStyles = makeStyles({
     },
 
     dialogContent: {
-        display: 'flex',
-        lineBreak: 'anywhere'
+        display: 'flex'
     },
     divider: {
         marginTop: '0.5rem',
@@ -47,9 +45,9 @@ const useStyles = makeStyles({
     }
 });
 
-type VoteValGroupConfirmProps = { lockAmount: string; validatorGroup?: string };
+type ActivateVoteValGroupConfirmProps = { validatorGroup?: string };
 
-const Confirm = ({ lockAmount, validatorGroup }: VoteValGroupConfirmProps): JSX.Element => {
+const Confirm = ({ validatorGroup }: ActivateVoteValGroupConfirmProps): JSX.Element => {
     const classes = useStyles();
     const [currentUser, setCurrentUser] = React.useState('');
 
@@ -71,8 +69,9 @@ const Confirm = ({ lockAmount, validatorGroup }: VoteValGroupConfirmProps): JSX.
                                     noWrap={false}
                                     color="textPrimary"
                                     gutterBottom>
-                                    You’re going to vote {lockAmount} CELO to {validatorGroup}, if
-                                    that is correct, please sign in your ledger device.
+                                    You’re going to activate your votes for Validator Group with
+                                    address {validatorGroup}. If that is correct, please sign in
+                                    your ledger device.
                                 </Typography>
                             </Grid>
                             <Grid item xs={12}>
@@ -122,31 +121,6 @@ const Confirm = ({ lockAmount, validatorGroup }: VoteValGroupConfirmProps): JSX.
                                         align="right"
                                         color="textPrimary">
                                         {validatorGroup}
-                                    </Typography>
-                                </Grid>
-                            </Grid>
-                            <Grid item xs={12}>
-                                <Divider className={classes.divider} />
-                            </Grid>
-                            <Grid container spacing={1} className={classes.item}>
-                                <Grid item xs={6}>
-                                    <Typography
-                                        variant="body2"
-                                        noWrap
-                                        className={classes.alignLeft}
-                                        align="left"
-                                        color="textPrimary">
-                                        Amount
-                                    </Typography>
-                                </Grid>
-                                <Grid item xs={6}>
-                                    <Typography
-                                        variant="body2"
-                                        noWrap
-                                        className={classes.alignRight}
-                                        align="right"
-                                        color="textPrimary">
-                                        {lockAmount} CELO
                                     </Typography>
                                 </Grid>
                             </Grid>