+++ components/ledger/validatorGroup/vote/Success.tsx
@@ -61,13 +61,10 @@ const useStyles = makeStyles({
     }
 });
 
-const Success = (): JSX.Element => {
-    const classes = useStyles();
-    const [open, setOpen] = React.useState(false);
+type SuccessProps = { txHash: string };
 
-    const handleClose = () => {
-        setOpen(false);
-    };
+const Success = ({ txHash }: SuccessProps): JSX.Element => {
+    const classes = useStyles();
 
     return (
         <>
@@ -101,14 +98,13 @@ const Success = (): JSX.Element => {
                                 className={classes.controlButton}
                                 alignItems="center">
                                 <NavLink
-                                    href="/transactions"
+                                    href={`/transaction/${txHash}`}
                                     name={
                                         <Button
                                             variant="outlined"
                                             color="secondary"
                                             className={classes.proposalButtonLabel}
-                                            fullWidth={true}
-                                            onClick={handleClose}>
+                                            fullWidth={true}>
                                             <Typography variant="body2" noWrap>
                                                 View Transactions
                                             </Typography>